Update MirrorOp.download.recipe
Updated to use new URL from www.barco.com. Had to add a PathDeleter at the beginning to delete the recipe cache dir as Autopkg would error out during the URLDownloader with a curl error 22/HTTP 400 on subsequent runs if a DMG was previously downloaded.

Running curl against the source URL directly didn't generate those errors, even when overwriting a pre-existing output file, so I gave up and went for the sledgehammer option (PathDeleter). It's crude and means you get a 18MB download each time the recipe runs, but it works... Hopefully someone with more knowledge might find a curl option or something that means we don't have to do that.