Add automatic version detection to RUM.pkg recipe
Searches the executable for text in an embedded plist that provides the version info (rather than the user having to manually enter it). Also removes the redundant reference to DOWNLOAD_URL (it only needs to appear in the parent recipe, or, if my other PR is accepted, not at all).